chore(component): move `src/components/Tables/TopBlockedQueries.tsx` to `src/components/Tables/Domains/TopBlockedDomains.tsx`
  ## what
  - move `src/components/Tables/TopBlockedQueries.tsx` to `src/components/Tables/Domains/TopBlockedDomains.tsx`
  - update references
    - ./src/pages/index.tsx

  ## how

  ## why
  - to organize components better

  ## where
  - ./src/components/Tables/TopBlockedQueries.tsx -> src/components/Tables/Domains/TopBlockedDomains.tsx
  - ./src/pages/index.tsx

  ## usage
